float16 broken in llvmlite use float64 for now

+++ test/test_assign.py
@@ -68,11 +68,11 @@ class TestAssign(unittest.TestCase):
     a = Tensor(np.arange(N*N, dtype=np.float32)).reshape(N,N)
     a.realize()
     oba1 = a.lazydata.output_buffer
-    a.assign(a.cast(dtypes.float16).realize())
+    a.assign(a.cast(dtypes.float64).realize())
     a.realize()
     oba2 = a.lazydata.output_buffer
     assert oba1 is None and oba2 is None
-    np.testing.assert_allclose(a.numpy(), np.arange(N*N,dtype=np.float16).reshape((N,N)))
+    np.testing.assert_allclose(a.numpy(), np.arange(N*N,dtype=np.float64).reshape((N,N)))
 
 if __name__ == "__main__":
   unittest.main()
